+++ udata_front/theme/gouvfr/assets/js/components/MarkdownEditor/ImageModal/ImageModal.stories.ts
@@ -1,5 +1,5 @@
 import { withActions } from '@storybook/addon-actions/decorator';
-import { expect } from '@storybook/jest';
+import { expect } from '@storybook/test';
 import { screen, userEvent, waitFor } from '@storybook/testing-library';
 import type { Meta, StoryObj } from '@storybook/vue3';
 import ImageModal, {ImageModalForm, type ImageModalProps} from './ImageModal.vue';