+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fix datastructure display in ResourceAccordion [#389](https://github.com/etalab/udata-front/pull/389)
 - Order Organization's reuses by publishing date [#390](https://github.com/etalab/udata-front/pull/390)
 - Fix failing captchEtat responses [#392](https://github.com/etalab/udata-front/pull/392)
+- Add beta admin member page [#374](https://github.com/etalab/udata-front/pull/374)
 
 ## 3.5.4 (2024-03-20)